pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protein-design-hub"
version = "0.1.0"
description = "Unified protein structure prediction and evaluation hub"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Protein Design Hub Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "biopython>=1.81",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.28.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pdhub = "protein_design_hub.cli.main:app"

[project.urls]
Homepage = "https://github.com/recep2244/pdhub"
Documentation = "https://github.com/recep2244/pdhub#readme"
Repository = "https://github.com/recep2244/pdhub"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
protein_design_hub = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
